requirements.txt
Flask-RESTful
Flask-SQLAlchemy
alembic
Flask-Migrate